Renamed edit-card.js to edit-button.js for better understanding of what component does
