+++ styles.css
@@ -34,4 +34,5 @@
   justify-content: flex-end;
   padding-right: 2px;
   padding-bottom: 2px;
+  cursor: pointer;
 }
